Move 4 manage_action_container tests to test_containers
