+++ components/tools/OmeroWeb/test/integration/test_csrf.py
@@ -24,7 +24,7 @@ working correctly.
 
 import omero
 import omero.clients
-from omero.rtypes import rstring, rtime
+from omero.rtypes import rstring
 from weblibrary import IWebTest
 from weblibrary import _csrf_post_response, _post_response
 from weblibrary import _csrf_get_response, _get_response
@@ -100,47 +100,6 @@ class TestCsrf(IWebTest):
         data = {
             'username': "omename",
             'email': "email"
-        }
-        _post_response(self.django_client, request_url, data)
-        _csrf_post_response(self.django_client, request_url, data)
-
-    def test_add_and_rename_container(self):
-
-        # Add project
-        request_url = reverse("manage_action_containers",
-                              args=["addnewcontainer"])
-        data = {
-            'folder_type': 'project',
-            'name': 'foobar'
-        }
-        _post_response(self.django_client, request_url, data)
-        response = _csrf_post_response(self.django_client, request_url, data)
-        pid = json.loads(response.content).get("id")
-
-        # Add dataset to the project
-        request_url = reverse("manage_action_containers",
-                              args=["addnewcontainer", "project", pid])
-        data = {
-            'folder_type': 'dataset',
-            'name': 'foobar'
-        }
-        _post_response(self.django_client, request_url, data)
-        _csrf_post_response(self.django_client, request_url, data)
-
-        # Rename project
-        request_url = reverse("manage_action_containers",
-                              args=["savename", "project", pid])
-        data = {
-            'name': 'anotherfoobar'
-        }
-        _post_response(self.django_client, request_url, data)
-        _csrf_post_response(self.django_client, request_url, data)
-
-        # Change project description
-        request_url = reverse("manage_action_containers",
-                              args=["savedescription", "project", pid])
-        data = {
-            'description': 'anotherfoobar'
         }
         _post_response(self.django_client, request_url, data)
         _csrf_post_response(self.django_client, request_url, data)
@@ -253,92 +212,6 @@ class TestCsrf(IWebTest):
 
         # Remove file, see remove tag,
         # http://localhost/webclient/action/remove/[comment|tag|file]/ID/
-
-    def test_paste_move_remove_deletamany_image(self):
-
-        # Add dataset
-        request_url = reverse("manage_action_containers",
-                              args=["addnewcontainer"])
-        data = {
-            'folder_type': 'dataset',
-            'name': 'foobar'
-        }
-        _post_response(self.django_client, request_url, data)
-        response = _csrf_post_response(self.django_client, request_url, data)
-        did = json.loads(response.content).get("id")
-
-        img = self.image_with_channels()
-
-        # Link image to Dataset
-        request_url = reverse("api_links")
-        data = {
-            'dataset': {did: {'image': [img.id.val]}}
-        }
-
-        _post_response(self.django_client, request_url, data)
-        _csrf_post_response(self.django_client,
-                            request_url,
-                            json.dumps(data),
-                            content_type="application/json")
-
-        # Unlink image from Dataset
-        request_url = reverse("api_links")
-        data = {
-            'dataset': {did: {'image': [img.id.val]}}
-        }
-        _delete_response(self.django_client, request_url, data)
-        response = _csrf_delete_response(self.django_client,
-                                         request_url,
-                                         json.dumps(data),
-                                         content_type="application/json")
-        # Response will contain remaining links from image (see test_links.py)
-        response = json.loads(response.content)
-        assert response == {"success": True}
-
-    def test_create_share(self):
-
-        img = self.image_with_channels()
-        request_url = reverse("manage_action_containers",
-                              args=["add", "share"])
-        data = {
-            'enable': 'on',
-            'image': img.id.val,
-            'members': self.user.id.val,
-            'message': 'foobar'
-        }
-
-        _post_response(self.django_client, request_url, data)
-        _csrf_post_response(self.django_client, request_url, data)
-
-    def test_edit_share(self):
-
-        # create images
-        images = [self.createTestImage(session=self.sf),
-                  self.createTestImage(session=self.sf)]
-
-        sid = self.sf.getShareService().createShare(
-            "foobar", rtime(None), images, [self.user], [], True)
-
-        request_url = reverse("manage_action_containers",
-                              args=["save", "share", sid])
-
-        data = {
-            'enable': 'on',
-            'image': [i.id.val for i in images],
-            'members': self.user.id.val,
-            'message': 'another foobar'
-        }
-        _post_response(self.django_client, request_url, data)
-        _csrf_post_response(self.django_client, request_url, data)
-
-        # remove image from share
-        request_url = reverse("manage_action_containers",
-                              args=["removefromshare", "share", sid])
-        data = {
-            'source': images[1].id.val,
-        }
-        _post_response(self.django_client, request_url, data)
-        _csrf_post_response(self.django_client, request_url, data)
 
     def test_edit_channel_names(self):
